mantap

+++ ais/db/migrate/20230625044755_create_blockchain_syncs.rb
@@ -2,7 +2,7 @@ class CreateBlockchainSyncs < ActiveRecord::Migration[6.1]
   def change
     create_table :blockchain_syncs do |t|
       t.string :syncable_type
-      t.integer :syncable_id
+      t.bigint :syncable_id
       t.string :status
       t.string :description
